stringstream write only variables of a fixed size, maybe will fix the op_target buffs crash... maybe not.

+++ trunk/EQEmuServer/common/eq_packet_structs.h
@@ -4421,7 +4421,7 @@ struct BuffIconEntry_Struct
 struct BuffIcon_Struct
 {
 	uint32 entity_id;
-	uint32 count;
+	uint16 count;
 	BuffIconEntry_Struct entries[0];
 };
